+++ Circles/Views/Composer/ImagePicker.swift
@@ -17,7 +17,7 @@ struct ImagePicker: UIViewControllerRepresentable {
         Coordinator(self)
     }
     
-    @Binding var selectedImage: UIImage?
+    //@Binding var selectedImage: UIImage?
     @Environment(\.presentationMode) private var presentationMode
     
     var sourceType: UIImagePickerController.SourceType = .photoLibrary
@@ -52,7 +52,7 @@ struct ImagePicker: UIViewControllerRepresentable {
         func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
      
             if let image = info[UIImagePickerController.InfoKey.originalImage] as? UIImage {
-                parent.selectedImage = image
+                //parent.selectedImage = image
                 parent.presentationMode.wrappedValue.dismiss()
 
                 // cvw: This is where we should call our "completion" closure to do whatever action we're supposed to do with the new image