+++ Circles/Views/Circles/CircleOverviewCard.swift
@@ -48,8 +48,9 @@ struct CircleOverviewCard: View {
                         if let wall = space.wall {
                             let knockCount = wall.knockingMembers.count
                             if knockCount > 0 {
+                                let color = Color(light: .accentColor, dark: .white)
                                 Label("\(knockCount) requests for invitations", systemImage: "star.fill")
-                                    .foregroundColor(.accentColor)
+                                    .foregroundColor(color)
                             }
                         }
                     }