+++ Circles/Views/Groups/GroupInvitationsView.swift
@@ -27,6 +27,7 @@ struct GroupInvitationsView: View {
                     }
                 }
             }
+            .frame(maxWidth: 500)
         }
         .navigationTitle(Text("Group Invitations"))
     }