+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# main
+* Update privacy policy URL
+
 # v0.9.6
 * Fix decryption errors for messages posted while the user is invited
 * Add a debug details section to the room settings