+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Bugfix: Set displayname and profile photo on the server at setup time
 * Bugfix: Enable taking new photo for your profile
 * Bugfix: Enable posting new photo from the camera
+* Make deletion safe again: Don't re-use already uploaded media
 * Update privacy policy URL
 
 # v0.9.6