+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # main
+* Update your profile picture everywhere when you choose a new one
 * Get rid of the ugly password nag screen
 * Speed up creating the Circles space hierarchy
 * Add the ability to share or delete individual photos