+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # main
+* Make requests for invitations more visible in dark mode
 * Remove photo gallery from our list when we leave
 * Automatically load photos when scrolling through a gallery
 * Prevent reactions from getting squished